+++ reus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.6"
+__version__ = "1.0.7"
 
 from .fbref import *
 from .transfermarkt import *